+++ src/Products/MailHost/MailHost.py
@@ -93,6 +93,7 @@ class MailBase(Implicit, Item, RoleManager):
     """a mailhost...?"""
 
     meta_type = 'Mail Host'
+    zmi_icon = 'icon-mail'
     manage = manage_main = DTMLFile('dtml/manageMailHost', globals())
     manage_main._setName('manage_main')
     index_html = None